fix(money): widen 4 accumulated price_cents i32 → i64 violations + migration 069
⚠️ Consolidates 4 CLAUDE.md "Money/cents" hard-rule violations flagged across
R12-D, R13-B, R15-D into one delivered fix.

CLAUDE.md (project root + ~/.claude/CLAUDE.md):
> Money is i64 / Postgres BIGINT ONLY at every layer. i32 / PG INTEGER caps
> at $21,474,836.47.

# Migration
- api/migrations/069_widen_money_cents_to_bigint.sql (next number; flat-file
  style matching 067, 068)
- Defensive: BIGINT for any column that wasn't already widened; documented
  no-op for phantom columns

# Rust files changed (7 fixes across 5 files)

1. api/src/models/course.rs:235
   UserCourseEnrollment.price_paid_cents: Option<i32> → Option<i64>
   (R12-D-flagged)

2. api/src/models/indicator.rs:137
   UserIndicatorOwnership.price_paid_cents: Option<i32> → Option<i64>
   (R12-D-flagged)

3. api/src/models/indicator_enhanced.rs:54,189,212,380
   FOUR fields widened:
   - IndicatorEnhanced.price_cents
   - CreateIndicatorRequest.price_cents
   - UpdateIndicatorRequest.price_cents
   - IndicatorResponse.price_cents
   (R15-D-flagged. Was silently narrowing-on-read since migration 061 made
   indicators_enhanced.price_cents BIGINT — Rust was reading BIGINT into i32.)

4. api/src/routes/admin_courses/analytics_pricing.rs:222
   CourseForPriceChange.price_cents: i32 → i64

5. api/src/routes/admin_courses/analytics_pricing.rs:283
   REMOVED `as i32` downcast on bind — change_course_price was writing
   `.bind(input.amount_cents as i32)` into a BIGINT column. Silent truncation
   above $21M. (R13-B-flagged.)

6. api/src/services/search.rs:19
   SearchableCourse.price_cents: i32 → i64 (cleanup; same struct used in
   search indexer)

# Key findings (documented in migration header)

- `user_course_enrollments.price_paid_cents` and `user_indicator_access.
  price_paid_cents` DON'T EXIST in any migration or current schema — verified
  with `grep -rEn "price_paid_cents" api/migrations/` (zero hits). The
  R12-D-flagged Rust types referenced phantom columns. Only widened Rust
  types so any future column-add lands as BIGINT.
- `courses.price_cents` and `indicators.price_cents` were already BIGINT after
  migration 068; migration 069 defensively re-asserts.
- `indicators_enhanced.price_cents` was BIGINT since migration 061; Rust was
  carrying Option<i32> against it (silent narrowing on read).

# Tests

The R12-D test `user_course_enrollment_ids_are_i64_documents_price_paid_landmine`
that PINNED the i32 landmine still passes — its inline comment explicitly
anticipated the widening would not break it.

# No sqlx::query_as! macros require live DB to compile
All affected paths use the runtime-checked non-macro form.

# Gates green (Rust MCP)
- cargo-check --all-targets PASS
- cargo-fmt --check PASS
- cargo-test (8 test crates, 69 tests total) all PASS:
  stripe_test, utils_test, payments_test, member_courses_test,
  admin_indicators_test, courses_admin_test, member_indicators_test, checkout_test
- cargo-clippy: 3 pre-existing -D warnings lints in unrelated files
  (cache/service.rs, domain/content/entity.rs); none in this PR's files;
  none introduced by this PR.

+++ api/src/models/course.rs
@@ -232,7 +232,12 @@ pub struct UserCourseEnrollment {
     pub access_expires_at: Option<NaiveDateTime>,
     pub is_lifetime_access: Option<bool>,
     pub order_id: Option<i64>,
-    pub price_paid_cents: Option<i32>,
+    // CLAUDE.md money rule: every *_cents value is i64 / BIGINT end-to-end.
+    // Widened from Option<i32> (audit R17-D, 2026-05-20). Note: this struct
+    // maps a legacy schema and is not currently bound to a live column;
+    // widening here keeps the type contract aligned with the rule for the
+    // day a `price_paid_cents` column is added.
+    pub price_paid_cents: Option<i64>,
     pub certificate_issued: Option<bool>,
     pub certificate_url: Option<String>,
     pub certificate_issued_at: Option<NaiveDateTime>,
